refactor: 어노테이션으로 인스턴스의 scope를 관리할 수 있도록 어노테이션 정의
