--- alsonglibrary2/src/main/java/com/example/alsonglibrary2/di/AlsongQualifier.kt
@@ -1,3 +1,0 @@
-package com.example.alsonglibrary2.di
-
-annotation class AlsongQualifier
