+++ alsonglibrary2/src/main/java/com/example/alsonglibrary2/di/anotations/AlsongQualifier.kt
@@ -1,0 +1,3 @@
+package com.example.alsonglibrary2.di.anotations
+
+annotation class AlsongQualifier
