+++ alsonglibrary2/src/main/java/com/example/alsonglibrary2/di/anotations/FieldInject.kt
@@ -1,4 +1,4 @@
-package com.example.alsonglibrary2.di
+package com.example.alsonglibrary2.di.anotations
 
 @Target(AnnotationTarget.PROPERTY)
 @Retention(AnnotationRetention.RUNTIME)
