+++ alsonglibrary2/src/main/java/com/example/alsonglibrary2/di/anotations/ViewModelScope.kt
@@ -1,0 +1,4 @@
+package com.example.alsonglibrary2.di.anotations
+
+@Target(AnnotationTarget.PROPERTY)
+annotation class ViewModelScope
