+++ alsonglibrary2/src/test/java/com/example/alsonglibrary2/fixtures/viewmodel/FieldInjectedWithTwoQualifierFakeViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.alsonglibrary2.fixtures.viewmodel
 
 import androidx.lifecycle.ViewModel
-import com.example.alsonglibrary2.di.FieldInject
+import com.example.alsonglibrary2.di.anotations.FieldInject
 import com.example.alsonglibrary2.fixtures.DefaultFakeRepository1
 import com.example.alsonglibrary2.fixtures.DefaultFakeRepository2
 import com.example.alsonglibrary2.fixtures.repository.FakeRepository
