+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping
 
 import android.app.Application
+import android.content.Context
 import com.example.alsonglibrary2.di.AutoDIManager
 import com.example.alsonglibrary2.di.AutoDIManager.createAutoDIInstance
 import com.example.alsonglibrary2.di.AutoDIManager.registerDependency
@@ -9,6 +10,7 @@ import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.DefaultCartRepository
 import woowacourse.shopping.data.InMemoryCartRepository
 import woowacourse.shopping.data.ShoppingDatabase
+import woowacourse.shopping.ui.cart.DateFormatter
 import woowacourse.shopping.ui.util.QualifierDependencyProvider
 
 class ShoppingApplication : Application() {
@@ -18,6 +20,7 @@ class ShoppingApplication : Application() {
         super.onCreate()
         registerDependencies()
         AutoDIManager.qualifierDependencyProvider = QualifierDependencyProvider
+        shoppingAppContext = this
     }
 
     private fun registerDependencies() {
@@ -26,7 +29,9 @@ class ShoppingApplication : Application() {
     }
 
     companion object {
+        lateinit var shoppingAppContext: Context
         val defaultCartRepository by lazy { createAutoDIInstance<DefaultCartRepository>() }
         val inMemoryCartRepository by lazy { createAutoDIInstance<InMemoryCartRepository>() }
+        val dateFormatter: DateFormatter by lazy { DateFormatter(shoppingAppContext) }
     }
 }
