+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.alsonglibrary2.di.FieldInject
+import com.example.alsonglibrary2.di.anotations.FieldInject
 import kotlinx.coroutines.launch
 import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.model.Product
