+++ app/src/main/java/woowacourse/shopping/ui/util/QualifierDependencyProvider.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.ui.util
 
-import com.example.alsonglibrary2.di.AlsongQualifier
 import com.example.alsonglibrary2.di.QualifierDependencyProvider
+import com.example.alsonglibrary2.di.anotations.AlsongQualifier
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.CartRepository
